Rename packages.py to repository.py, PackageFinder to RepoPath.
